# api-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Website for brewing potions and elixirs.",
        "title": "AlchemicalLab",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/brew": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brewing"
                ],
                "summary": "Started brewing a potion",
                "parameters": [
                    {
                        "description": "Brew",
                        "name": "brew",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.JobUUIDDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "recipe not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "ingredient not enough",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/brew/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brewing"
                ],
                "summary": "Started to brewing potion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.JobStatusDTO"
                        }
                    },
                    "404": {
                        "description": "job not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/ingredients": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Create and add new ingredients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.IngredientResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Create and add new ingredients",
                "parameters": [
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientResponseDTO"
                        }
                    },
                    "409": {
                        "description": "already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ingredients/{id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Add quantity to exist ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "recipe",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quantity to add",
                        "name": "quantity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientAddDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "ingredient not found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/recipes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Get all recipes",
                "parameters": [
                    {
                        "description": "Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.IngredientResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Create new recipe",
                "parameters": [
                    {
                        "description": "Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RecipeDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.IngredientResponseDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "recipe already exists",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.IngredientAddDTO": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dto.IngredientDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dto.IngredientResponseDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dto.JobDTO": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "recipe_id": {
                    "type": "integer"
                }
            }
        },
        "dto.JobStatusDTO": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.JobUUIDDTO": {
            "type": "object",
            "properties": {
                "job_uuid": {
                    "type": "string"
                }
            }
        },
        "dto.RecipeDTO": {
            "type": "object",
            "required": [
                "ingredients",
                "name"
            ],
            "properties": {
                "brewing_time_seconds": {
                    "type": "integer"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "ingredients": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.RecipeIngredientsDTO"
                    }
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
        "dto.RecipeIngredientsDTO": {
            "type": "object",
            "required": [
                "ingredient_id",
                "quantity_needed"
            ],
            "properties": {
                "ingredient_id": {
                    "type": "integer"
                },
                "quantity_needed": {
                    "type": "integer"
                }
            }
        }
    }
}